Core: Add AbstractTable abstraction
AbstractTable is a class which represents everything that has rows and
columns. It is read only, because not all SQL tables are writable, and
this would add much more complexity to the API - every implementer would
need to implement all of insert(), delete(), add_column() etc.

The only class that implements AbstractTable is Table for now. The
SELECT statement is ported to AbstractTable as a proof of concept.