GUI: Implement DatabaseClient abstraction
This commit moves all the engine-specific code into implementations of
the DatabaseConnection class.

The GUI is also reworked so that you always need to "open a connection"
even if you do this for client-side EssaDB engine. The GUI asks user for
connection type, then a connection type-specific dialog is opened (like
for MySQL before).