Core: Store dates in wall-clock time
This assumes that all dates are in the same (unspecified) timezone and
calendar over the entire database. It makes dates specified in ISO8601
format appear as entered (they were sometimes offsetted by 1 day for
some reason).

This should fix CI failing on dates because of being in a different
timezone.