Storage: Update prev row iterator on row removal
Previously we were leaving it as is, pointing to the freed row, which
caused EDBFile::remove() to be called with wrong prev row and
leaving first_row_ptr in header pointing to the second (deleted) row.

Now we are updating EDBRelationIterator prev row pointer just after it
was deleted. This fixes crash that was occuring when selecting from
table that was cleared by 'DELETE FROM' just after inserting. A test is
added for this case.