Storage/EDB: Fix access use after invalidation on insertion
When writing row, we created Access (with the old address),
invalidated it in Serializer::write_row() (which may allocate on heap,
and therefore expand & remap the file) and then keeping it to be flushed
at the end of the function (to that old address). This commit moves all
row operations to a serializer block and do them only after calling
Serializer::write_row to ensure the new address is always used.

Note: Linux returning a address 1 page before didn't make debugging
easier :/